add Access History

+++ UrlShortener.Persistence/Interfaces/IGenericRepository.cs
@@ -5,7 +5,7 @@ namespace UrlShortener.Persistence.Interfaces
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> Get(int id);
+        Task<T> Get(long id);
         Task<IEnumerable<T>> GetAll();
         Task Add(T entity);
         void Delete(T entity);
