+++ UrlShortener.Persistence/Repositories/GenericRepository.cs
@@ -14,7 +14,7 @@ namespace UrlShortener.Persistence.Repositories
             _context = context;
         }
 
-        public async Task<T> Get(int id)
+        public async Task<T> Get(long id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
